Clean up HTML Encoding return output
The regex used previously was a bit too naive of an implementation in my
opinion, and also would cause issues if users, for instance, wished to
log raw XML to the console for troubleshooting (regex would have
stripped out the XML tags altogether).

The more correct approach in my opinion is to HTML escape the
characters. There is an HtmlEncode function that we could use, however
it would require importing System.Web, which I feel is a bit overkill
when we're simply escaping 4 problem characters.

There was also an issue with the previous commit that left an unused
branch of code in Output that still insisted on using \n. I removed the
unused branch as it is safer and cleaner to just go with always using
HTML since we're using a <div> instead of <textarea>.