Use Color32 consistently across Server.cs and Console.cs
Previously, Server.cs was using Color to represent the end user selected
colors, while Console.cs was representing this using a Color32. There
were no problems with converting between these of course, but given the
nature of this asset, we don't need the fidelity of color that Color
would provide.

The benifits of Color32 is the color data is stored in 4 bytes, whereas
Color is stored in 4 floats. Color has some static defaults that can be
used, however these can still be cast fine to a Color32 with no problem.
And you can still set user defined presets to the object as well.

As such, I updated the MessageColors struct to use Color32 to keep this
consistent.