Remove color mapping in server
This was actually a bit overengineered in the previous implementation,
after feedback I agree it's much simpler to just go with a direct change
to the log API to specify the color there, instead of using a string
mapping approach. I went with that approach previously as I thought that
it may be nice to have a "message log type" concept, but that's actually
a pretty convoluted approach.

Changes:
- Remove changes to Server.cs
- Remove the mapping structure from Console.cs
- Update Log(string, string) to be Log(string, Color32) instead.